+++ Activities/GameActivity.cpp
@@ -387,7 +387,7 @@ bool GameActivity::IsBuyGUIVisible(int which) const {
 // Description:     Forces the this to focus player control to a specific Actor for a
 //                  specific team. OWNERSHIP IS NOT TRANSFERRED!
 
-bool GameActivity::SwitchToActor(Actor *pActor, short player, short team)
+bool GameActivity::SwitchToActor(Actor *pActor, int player, int team)
 {
     // Computer players don't focus on any Actor
     if (!m_IsHuman[player])
@@ -410,7 +410,7 @@ bool GameActivity::SwitchToActor(Actor *pActor, short player, short team)
 // Description:     Forces the this to focus player control to the next Actor of a
 //                  specific team, other than the current one focused on.
 
-void GameActivity::SwitchToNextActor(short player, short team, Actor *pSkip)
+void GameActivity::SwitchToNextActor(int player, int team, Actor *pSkip)
 {
     // Play the disabling animation when the actor swtiched, for easy ID of currently controlled actor
     m_pPieMenu[player]->DisableAnim();
@@ -429,7 +429,7 @@ void GameActivity::SwitchToNextActor(short player, short team, Actor *pSkip)
 // Description:     Forces this to focus player control to the previous Actor of a
 //                  specific team, other than the current one focused on.
 
-void GameActivity::SwitchToPrevActor(short player, short team, Actor *pSkip)
+void GameActivity::SwitchToPrevActor(int player, int team, Actor *pSkip)
 {
     // Play the disabling animation when the actor swtiched, for easy ID of currently controlled actor
     m_pPieMenu[player]->DisableAnim();
